requirements.txt
langchain_community.utilities
langchain_ollama
langchain_core.prompts
streamlit
mysql-connector-python